+++ src/main/java/main/java/Duke.java
@@ -1,0 +1,96 @@
+package main.java;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+import java.util.Scanner;
+
+import java.io.File;
+import java.nio.file.Files;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.FileNotFoundException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.LocalDateTime;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.text.ParseException;
+
+
+// for level 9, use contains and check each task, appending to the temp array result
+
+public class Duke {
+
+    private Storage storage;
+    private TaskList tasks;
+    private Ui ui;
+
+    public Duke(String filePath) {
+        ui = new Ui();
+        storage = new Storage(filePath);
+        try {
+            tasks = new TaskList(storage.load());
+        } catch (FileNotFoundException e) {
+            ui.showLoadingError();
+            tasks = new TaskList();
+        }
+    }
+
+
+    private static void printFileContents(String filePath) throws FileNotFoundException {
+        File f = new File(filePath); // create a File for the given file path
+        Scanner s = new Scanner(f); // create a Scanner using the File as the source
+        while (s.hasNext()) {
+            System.out.println(s.nextLine());
+        }
+    }
+
+
+    // taken from Week 3 topics
+
+
+
+    public void run() {
+        try {
+            File file = new File("data/duke.txt");
+            System.out.println("Current Tasks:");
+            printFileContents("data/duke.txt");
+            // no need to get tasks here, alr done in load at constructor
+//            System.out.println(this.tasks);
+        } catch (FileNotFoundException e) {
+            System.out.println("File not found, added file");
+        }
+
+        // put as attributes in ui
+        ui.showWelcomeMessage();
+
+        Scanner scanner = new Scanner(System.in);
+
+        ArrayList<Boolean> isOpen = new ArrayList<>();
+        isOpen.add(true);
+
+        try {
+            while (scanner.hasNextLine() && isOpen.get(0)) {
+                String input = scanner.nextLine();
+                String command = this.ui.getUserCommand(input);
+                Parser.parse(command, input, tasks, isOpen, scanner);
+//                System.out.println(isOpen.get(0));
+            }
+        } catch (IllegalStateException e) {
+            // just catching error
+        }
+        scanner.close();
+    }
+
+    // dealt with getting from file initially and saving, but need to call function for rewrite somewhere
+    public static void main(String[] args) throws DukeException{
+        new Duke("data/tasks.txt").run();
+    }
+}
+
+
