+++ src/main/java/main/java/DukeException.java
@@ -1,3 +1,5 @@
+package main.java;
+
 public class DukeException extends Exception{
 
     public DukeException(String message) {
