+++ src/main/java/main/java/Event.java
@@ -1,3 +1,5 @@
+package main.java;
+
 public class Event extends Task {
 
     protected String dueTime;
