+++ src/main/java/main/java/Task.java
@@ -1,3 +1,5 @@
+package main.java;
+
 public class Task {
     protected String description;
     protected boolean isDone;
